Separated Session routes/tests and Graph route/tests
